refactoring export to be a mixin
